pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = ""
authors = [
    {name = "phi.nguyen@orientsoftware.com",email = "phi.nguyen@orientsoftware.com"}
]
readme = "README.md"
requires-python = "==3.12.7"
dependencies = [
    "jupyter (==1.1.1)",
    "streamlit (==1.41.1)",
    "python-dotenv (==1.0.1)",
    "openai (==1.59.5)",
    "loguru (==0.7.3)",
    "qdrant-client (>=1.12.2,<2.0.0)",
    "pydantic (>=2.10.5,<3.0.0)",
    "llama-index-core (>=0.12.10.post1,<0.13.0)",
    "llama-index-utils-workflow (>=0.3.0,<0.4.0)",
    "llama-parse (>=0.5.19,<0.6.0)",
    "tiktoken (>=0.8.0,<0.9.0)",
    "llama-index-vector-stores-qdrant (>=0.4.2,<0.5.0)",
    "llama-index-embeddings-openai (>=0.3.1,<0.4.0)",
    "llama-index-llms-openai (>=0.3.13,<0.4.0)",
    "llama-index-agent-openai (>=0.4.1,<0.5.0)",
    "python-pptx (>=1.0.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
